+++ xclim/core/formatting.py
@@ -567,6 +567,8 @@ def _gen_parameters_section(parameters: dict, allowed_periods: list[str] = None)
             defstr = ""
         elif param.default is not _empty:
             defstr = f"Default : {param.default}. "
+        else:
+            defstr = "Required. "
         if "choices" in param:
             annotstr = str(param.choices)
         else: